cosmetics

+++ applications/nester-test/main_nester_test.cpp
@@ -1,29 +1,17 @@
 #include <fstream>
 #include <list>
-
 #include "gtest/gtest.h"
-
 #include "core/utils/utils.h"
 #include "core/buffers/rgb24/abstractPainter.h"
 #include "core/buffers/rgb24/rgb24Buffer.h"
-
 #include "core/fileformats/svgLoader.h"
 #include <core/reflection/commandLineSetter.h>
-
 #include <algorithm>
 #include <core/buffers/bufferFactory.h>
 #define EPSIL 0.00001
-
 #include "../nester/nester.h"
-
 using namespace corecvs;
 using namespace std;
-
-
-//for some reason header do not see list
-
-
-
 //bool testDisjoint(std::vector<Polygon> &inputPolygons)
 //{
 //    std::vector<ConvexPolygon> cpl;
@@ -47,10 +35,6 @@ using namespace std;
 //    }
 
 //}
-
-
-
-
 TEST(Nester, twoRectanges)
 {
     Rectangled area (0, 0, 200, 100);
@@ -59,82 +43,49 @@ TEST(Nester, twoRectanges)
         Polygon::RegularPolygon(6, Vector2dd::Zero(), 20, 0),
         Polygon::RegularPolygon(7, Vector2dd::Zero(), 20, 0)
     };
-
     drawSvgPolygons(inpList, area.height(), area.width(), "in.svg");
 
     bottomLeftPlacementProtected(inpList, area, 3);
-
-
     drawSvgPolygons(inpList, area.height(), area.width(), "out.svg");
-
-
-
 }
 
 TEST(Nester, manyRectanges)
 {
     Rectangled area (0, 0, 500, 500); //1000, 1000 46
-
     list<Polygon> inpList;
-
     Vector2dd A(0,0);
     Vector2dd B(2, 10);
     Vector2dd C(20, 0);
-
-
     Polygon someTriangle = {A,B,C};
-
-
-
-
-    for ( int i = 0; i < 30; i++)
+    for (size_t i = 0; i < 30; i++)
     {
         auto Pol = Polygon::RegularPolygon(4, Vector2dd::Zero(), 20, degToRad(i));
         inpList.push_back(Pol);
     }
-
-    for ( int i = 0; i < 30; i++)
+    for (size_t i = 0; i < 30; i++)
     {
         auto Pol = Polygon::RegularPolygon(7, Vector2dd::Zero(), 20, degToRad(i));
         inpList.push_back(Pol);
     }
-
-    for ( int i = 0; i < 30; i++)
+    for (size_t i = 0; i < 30; i++)
     {
         auto Pol = Polygon::RegularPolygon(8, Vector2dd::Zero(), 20, degToRad(i));
         inpList.push_back(Pol);
     }
-
-
-    for(int i = 0; i < 20; ++i)
+    for (size_t i = 0; i < 20; ++i)
     {
         inpList.push_back(someTriangle);
     }
-
-
-
-
     drawSvgPolygons(inpList, area.height(), area.width(), "in1.svg");
 
     bottomLeftPlacementProtected(inpList, area, 4);
-
     drawSvgPolygons(inpList, area.height(), area.width(), "out1.svg");
-
-
 }
 
 TEST(Nester, nfp) //both figures must be right-oriented, or use doClockOrientation(Polygon)
 {
-
-
-
-
-
     Polygon nfpTestB = {{4,2}, {2,2}, {1,3}, {1,5}, {5, 3.5}};//ClockOrintated
     Polygon nfpTestA = {{4, 0}, {3, -0.5}, {2, -1}, {2, 1}}; //ClockOrintated
-
-
-
     Polygon nfpTestResult1 = convexNFP(nfpTestA, nfpTestB);
 
     int indWhere = getTopRightIndex(nfpTestResult1);
@@ -146,11 +97,7 @@ TEST(Nester, nfp) //both figures must be right-oriented, or use doClockOrientati
 
     list <Polygon> p = {nfpTestResult1, nfpTestA, nfpTestB, d};
     drawPolygons(p, 100, 100, "nfp1.bmp"); //does not draw last edge of nfp,  so d show that its all ok
-
-
-
 }
-
 
 int main(int argc, char **argv)
 {
